Add support for MENUITEMS and DISPLAY_PAGES_ON_MENU
With this commit, `.sidebar-nav` is renamed to `.sidebar-social`. Then a
new class `.sidebar-nav` is created and used for the menu items.  This
change more or less restores `.sidebar-nav` originally was in the
`hyde.css` file.

Since `DISPLAY_PAGES_ON_MENU` defaults to `True`, this commit will
modify the appearance of your site. To go back to the previous behavior,
add this to your pelicanconf.py:

    DISPLAY_PAGES_ON_MENU = False

References:
- <https://docs.getpelican.com/en/stable/settings.html#basic-settings>
- <https://docs.getpelican.com/en/stable/settings.html#MENUITEMS>